+++ amethyst_renderer/src/light.rs
@@ -51,7 +51,9 @@ impl From<DirectionalLight> for Light {
     }
 }
 
-/// A point light source.
+/// A point light source. Uses the `Transform` set of components for
+/// positioning, and requires a `GlobalTransform` component to be included 
+/// in rendering.
 ///
 /// Lighting calculations are based off of the Frostbite engine's lighting,
 /// which is explained in detail here in [this presentation][fb]. Below is
@@ -75,8 +77,6 @@ impl From<DirectionalLight> for Light {
 #[derive(Clone, ConstantBuffer, Debug, Deserialize, PartialEq, Serialize)]
 #[serde(default)]
 pub struct PointLight {
-    /// Location of the light source in three dimensional space.
-    pub center: [f32; 3], //TODO: Replace with a cgmath type when gfx version > 0.16
     /// Color of the light in RGBA8 format.
     pub color: Rgba,
     /// Brightness of the light source, in lumens.
@@ -91,7 +91,6 @@ pub struct PointLight {
 impl Default for PointLight {
     fn default() -> Self {
         PointLight {
-            center: [0.0, 0.0, 0.0],
             color: Rgba::default(),
             intensity: 10.0,
             radius: 10.0,