+++ examples/renderable/main.rs
@@ -5,21 +5,185 @@
 
 extern crate amethyst;
 
-use amethyst::assets::{Completion, HotReloadBundle, Loader, ProgressCounter};
-use amethyst::core::cgmath::{Array, Deg, Euler, Quaternion, Rad, Rotation, Rotation3, Vector3};
+use amethyst::assets::{Completion, Handle, HotReloadBundle, Prefab, PrefabLoader,
+                       PrefabLoaderSystem, ProgressCounter, RonFormat};
+use amethyst::core::cgmath::{Quaternion, Rad, Rotation, Rotation3};
 use amethyst::core::timing::Time;
-use amethyst::core::transform::{GlobalTransform, Transform, TransformBundle};
-use amethyst::ecs::prelude::{Component, Entity, Join, Read, ReadStorage, System, World,
-                             WriteExpect, WriteStorage};
-use amethyst::ecs::storage::NullStorage;
+use amethyst::core::transform::{Transform, TransformBundle};
+use amethyst::ecs::prelude::{Entity, Join, Read, ReadStorage, System, Write, WriteStorage};
 use amethyst::input::{get_key, is_close_requested, is_key, InputBundle};
-use amethyst::renderer::{AmbientColor, Camera, DirectionalLight, DrawShaded, Event, Light,
-                         Material, MaterialDefaults, MeshHandle, ObjFormat, PngFormat, PointLight,
-                         PosNormTex, Projection, Rgba, TextureFormat, VirtualKeyCode};
-use amethyst::ui::{Anchor, FontFormat, FontHandle, TtfFormat, UiBundle, UiCreator, UiFinder,
-                   UiLoaderSystem, UiText, UiTransform};
+use amethyst::renderer::{AmbientColor, Camera, DrawShaded, Event, Light, PosNormTex,
+                         VirtualKeyCode};
+use amethyst::ui::{UiBundle, UiCreator, UiFinder, UiText};
 use amethyst::utils::fps_counter::{FPSCounter, FPSCounterBundle};
+use amethyst::utils::scene::BasicScenePrefab;
 use amethyst::{Application, Error, GameData, GameDataBuilder, State, StateData, Trans};
+
+type MyPrefabData = BasicScenePrefab<Vec<PosNormTex>>;
+
+#[derive(Default)]
+struct Loading {
+    progress: ProgressCounter,
+    prefab: Option<Handle<Prefab<MyPrefabData>>>,
+}
+
+struct Example {
+    scene: Handle<Prefab<MyPrefabData>>,
+}
+
+impl<'a, 'b> State<GameData<'a, 'b>> for Loading {
+    fn on_start(&mut self, data: StateData<GameData>) {
+        self.prefab = Some(data.world.exec(|loader: PrefabLoader<MyPrefabData>| {
+            loader.load("prefab/renderable.ron", RonFormat, (), &mut self.progress)
+        }));
+
+        data.world.exec(|mut creator: UiCreator| {
+            creator.create("ui/fps.ron", &mut self.progress);
+            creator.create("ui/loading.ron", &mut self.progress);
+        });
+    }
+
+    fn handle_event(&mut self, _: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
+        if is_close_requested(&event) || is_key(&event, VirtualKeyCode::Escape) {
+            Trans::Quit
+        } else {
+            Trans::None
+        }
+    }
+
+    fn update(&mut self, data: StateData<GameData>) -> Trans<GameData<'a, 'b>> {
+        data.data.update(&data.world);
+        match self.progress.complete() {
+            Completion::Failed => {
+                println!("Failed loading assets: {:?}", self.progress.errors());
+                Trans::Quit
+            }
+            Completion::Complete => {
+                println!("Assets loaded, swapping state");
+                if let Some(entity) = data.world.exec(|finder: UiFinder| finder.find("loading")) {
+                    let _ = data.world.delete_entity(entity);
+                }
+                Trans::Switch(Box::new(Example {
+                    scene: self.prefab.as_ref().unwrap().clone(),
+                }))
+            }
+            Completion::Loading => Trans::None,
+        }
+    }
+}
+
+impl<'a, 'b> State<GameData<'a, 'b>> for Example {
+    fn on_start(&mut self, data: StateData<GameData>) {
+        let StateData { world, .. } = data;
+
+        world.create_entity().with(self.scene.clone()).build();
+    }
+
+    fn handle_event(&mut self, data: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
+        let w = data.world;
+        // Exit if user hits Escape or closes the window
+        if is_close_requested(&event) || is_key(&event, VirtualKeyCode::Escape) {
+            return Trans::Quit;
+        }
+        match get_key(&event) {
+            Some(VirtualKeyCode::R) => {
+                w.exec(|mut state: Write<DemoState>| {
+                    state.light_color = [0.8, 0.2, 0.2, 1.0];
+                });
+            }
+            Some(VirtualKeyCode::G) => {
+                w.exec(|mut state: Write<DemoState>| {
+                    state.light_color = [0.2, 0.8, 0.2, 1.0];
+                });
+            }
+            Some(VirtualKeyCode::B) => {
+                w.exec(|mut state: Write<DemoState>| {
+                    state.light_color = [0.2, 0.2, 0.8, 1.0];
+                });
+            }
+            Some(VirtualKeyCode::W) => {
+                w.exec(|mut state: Write<DemoState>| {
+                    state.light_color = [1.0, 1.0, 1.0, 1.0];
+                });
+            }
+            Some(VirtualKeyCode::A) => {
+                w.exec(
+                    |(mut state, mut color): (Write<DemoState>, Write<AmbientColor>)| {
+                        if state.ambient_light {
+                            state.ambient_light = false;
+                            color.0 = [0.0; 3].into();
+                        } else {
+                            state.ambient_light = true;
+                            color.0 = [0.01; 3].into();
+                        }
+                    },
+                );
+            }
+            Some(VirtualKeyCode::D) => {
+                w.exec(
+                    |(mut state, mut lights): (Write<DemoState>, WriteStorage<Light>)| {
+                        if state.directional_light {
+                            state.directional_light = false;
+                            for light in (&mut lights).join() {
+                                if let Light::Directional(ref mut d) = *light {
+                                    d.color = [0.0; 4].into();
+                                }
+                            }
+                        } else {
+                            state.directional_light = true;
+                            for light in (&mut lights).join() {
+                                if let Light::Directional(ref mut d) = *light {
+                                    d.color = [0.2; 4].into();
+                                }
+                            }
+                        }
+                    },
+                );
+            }
+            Some(VirtualKeyCode::P) => {
+                w.exec(|mut state: Write<DemoState>| {
+                    if state.point_light {
+                        state.point_light = false;
+                        state.light_color = [0.0; 4].into();
+                    } else {
+                        state.point_light = true;
+                        state.light_color = [1.0; 4].into();
+                    }
+                });
+            }
+            _ => (),
+        }
+        Trans::None
+    }
+
+    fn update(&mut self, data: StateData<GameData>) -> Trans<GameData<'a, 'b>> {
+        data.data.update(&data.world);
+        Trans::None
+    }
+}
+
+fn main() -> Result<(), Error> {
+    // Add our meshes directory to the asset loader.
+    let resources_directory = format!("{}/examples/assets/", env!("CARGO_MANIFEST_DIR"));
+
+    let display_config_path = format!(
+        "{}/examples/renderable/resources/display_config.ron",
+        env!("CARGO_MANIFEST_DIR")
+    );
+
+    let game_data = GameDataBuilder::default()
+        .with(PrefabLoaderSystem::<MyPrefabData>::default(), "", &[])
+        .with::<ExampleSystem>(ExampleSystem::default(), "example_system", &[])
+        .with_bundle(TransformBundle::new().with_dep(&["example_system"]))?
+        .with_bundle(UiBundle::<String, String>::new())?
+        .with_bundle(HotReloadBundle::default())?
+        .with_bundle(FPSCounterBundle::default())?
+        .with_basic_renderer(display_config_path, DrawShaded::<PosNormTex>::new(), true)?
+        .with_bundle(InputBundle::<String, String>::new())?;
+    let mut game = Application::build(resources_directory, Loading::default())?.build(game_data)?;
+    game.run();
+    Ok(())
+}
 
 struct DemoState {
     light_angle: f32,
@@ -28,6 +192,19 @@ struct DemoState {
     point_light: bool,
     directional_light: bool,
     camera_angle: f32,
+}
+
+impl Default for DemoState {
+    fn default() -> Self {
+        DemoState {
+            light_angle: 0.0,
+            light_color: [1.0; 4],
+            ambient_light: true,
+            point_light: true,
+            directional_light: true,
+            camera_angle: 0.0,
+        }
+    }
 }
 
 #[derive(Default)]
@@ -41,7 +218,7 @@ impl<'a> System<'a> for ExampleSystem {
         Read<'a, Time>,
         ReadStorage<'a, Camera>,
         WriteStorage<'a, Transform>,
-        WriteExpect<'a, DemoState>,
+        Write<'a, DemoState>,
         WriteStorage<'a, UiText>,
         Read<'a, FPSCounter>,
         UiFinder<'a>,
@@ -69,16 +246,16 @@ impl<'a> System<'a> for ExampleSystem {
             transform.rotation = (delta_rot * Quaternion::from(transform.rotation)).into();
         }
 
-        for point_light in (&mut lights).join().filter_map(|light| {
+        for (point_light, transform)  in (&mut lights, &mut transforms).join().filter_map(|(light, transform)| {
             if let Light::Point(ref mut point_light) = *light {
-                Some(point_light)
+                Some((point_light, transform))
             } else {
                 None
             }
         }) {
-            point_light.center[0] = light_orbit_radius * state.light_angle.cos();
-            point_light.center[1] = light_orbit_radius * state.light_angle.sin();
-            point_light.center[2] = light_z;
+            transform.translation.x = light_orbit_radius * state.light_angle.cos();
+            transform.translation.y = light_orbit_radius * state.light_angle.sin();
+            transform.translation.z = light_z;
 
             point_light.color = state.light_color.into();
         }
@@ -97,373 +274,4 @@ impl<'a> System<'a> for ExampleSystem {
             }
         }
     }
-}
-
-#[derive(Default)]
-struct Loading {
-    progress: Option<ProgressCounter>,
-}
-struct Example;
-
-#[derive(Default)]
-struct LoadTag;
-
-impl Component for LoadTag {
-    type Storage = NullStorage<Self>;
-}
-
-impl<'a, 'b> State<GameData<'a, 'b>> for Loading {
-    fn on_start(&mut self, data: StateData<GameData>) {
-        data.world.register::<LoadTag>();
-        let (mut progress, assets) = load_assets(&data.world);
-
-        data.world
-            .exec(|mut creator: UiCreator| creator.create("ui/renderable.ron", &mut progress));
-
-        self.progress = Some(progress);
-
-        data.world
-            .create_entity()
-            .with(UiTransform::new(
-                "loading".to_string(),
-                Anchor::Middle,
-                0.,
-                0.,
-                1.,
-                200.,
-                50.,
-                0,
-            ))
-            .with(UiText::new(
-                assets.font.clone(),
-                "Loading".to_string(),
-                [1.0, 1.0, 1.0, 1.0],
-                25.,
-            ))
-            .with(LoadTag)
-            .build();
-
-        data.world.add_resource(assets);
-        data.world.add_resource::<DemoState>(DemoState {
-            light_angle: 0.0,
-            light_color: [1.0; 4],
-            ambient_light: true,
-            point_light: true,
-            directional_light: true,
-            camera_angle: 0.0,
-        });
-    }
-
-    fn handle_event(&mut self, _: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
-        if is_close_requested(&event) || is_key(&event, VirtualKeyCode::Escape) {
-            Trans::Quit
-        } else {
-            Trans::None
-        }
-    }
-
-    fn update(&mut self, data: StateData<GameData>) -> Trans<GameData<'a, 'b>> {
-        data.data.update(&data.world);
-        match self.progress.as_ref().unwrap().complete() {
-            Completion::Failed => {
-                println!("Failed loading assets");
-                Trans::Quit
-            }
-            Completion::Complete => {
-                println!("Assets loaded, swapping state");
-                let entities = (
-                    &*data.world.entities(),
-                    &data.world.read_storage::<LoadTag>(),
-                ).join()
-                    .map(|(e, _)| e)
-                    .collect::<Vec<_>>();
-                if let Err(err) = data.world.delete_entities(&entities) {
-                    println!("Failed deleting load state entities: {}", err);
-                }
-                Trans::Switch(Box::new(Example))
-            }
-            Completion::Loading => Trans::None,
-        }
-    }
-}
-
-impl<'a, 'b> State<GameData<'a, 'b>> for Example {
-    fn on_start(&mut self, data: StateData<GameData>) {
-        let StateData { world, .. } = data;
-        initialise_camera(world);
-
-        let assets = world.read_resource::<Assets>().clone();
-
-        // Add teapot and lid to scene
-        for mesh in vec![assets.lid.clone(), assets.teapot.clone()] {
-            let mut trans = Transform::default();
-            trans.rotation = Quaternion::from(Euler::new(Deg(90.0), Deg(-90.0), Deg(0.0))).into();
-            trans.translation = Vector3::new(5.0, 5.0, 0.0);
-
-            world
-                .create_entity()
-                .with(mesh)
-                .with(assets.red.clone())
-                .with(trans)
-                .with(GlobalTransform::default())
-                .build();
-        }
-
-        // Add cube to scene
-        let mut trans = Transform::default();
-        trans.translation = Vector3::new(5.0, -5.0, 2.0);
-        trans.scale = [2.0; 3].into();
-
-        world
-            .create_entity()
-            .with(assets.cube.clone())
-            .with(assets.logo.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        // Add cone to scene
-        let mut trans = Transform::default();
-        trans.translation = Vector3::new(-5.0, 5.0, 0.0);
-        trans.scale = [2.0; 3].into();
-
-        world
-            .create_entity()
-            .with(assets.cone.clone())
-            .with(assets.white.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        // Add custom cube object to scene
-        let mut trans = Transform::default();
-        trans.translation = Vector3::new(-5.0, -5.0, 1.0);
-        world
-            .create_entity()
-            .with(assets.cube.clone())
-            .with(assets.red.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        // Create base rectangle as floor
-        let mut trans = Transform::default();
-        trans.scale = Vector3::from_value(10.);
-
-        world
-            .create_entity()
-            .with(assets.rectangle.clone())
-            .with(assets.white.clone())
-            .with(trans)
-            .with(GlobalTransform::default())
-            .build();
-
-        let light: Light = PointLight {
-            color: [1.0, 1.0, 0.0].into(),
-            intensity: 50.0,
-            ..PointLight::default()
-        }.into();
-
-        // Add lights to scene
-        world.create_entity().with(light).build();
-
-        let light: Light = DirectionalLight {
-            color: [0.2; 4].into(),
-            direction: [-1.0; 3],
-        }.into();
-
-        world.create_entity().with(light).build();
-        world.add_resource(AmbientColor(Rgba::from([0.01; 3])));
-    }
-
-    fn handle_event(&mut self, data: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
-        let w = data.world;
-        // Exit if user hits Escape or closes the window
-        let mut state = w.write_resource::<DemoState>();
-
-        if is_close_requested(&event) || is_key(&event, VirtualKeyCode::Escape) {
-            return Trans::Quit;
-        }
-        match get_key(&event) {
-            Some(VirtualKeyCode::R) => {
-                state.light_color = [0.8, 0.2, 0.2, 1.0];
-            }
-            Some(VirtualKeyCode::G) => {
-                state.light_color = [0.2, 0.8, 0.2, 1.0];
-            }
-            Some(VirtualKeyCode::B) => {
-                state.light_color = [0.2, 0.2, 0.8, 1.0];
-            }
-            Some(VirtualKeyCode::W) => {
-                state.light_color = [1.0, 1.0, 1.0, 1.0];
-            }
-            Some(VirtualKeyCode::A) => {
-                let mut color = w.write_resource::<AmbientColor>();
-                if state.ambient_light {
-                    state.ambient_light = false;
-                    color.0 = [0.0; 3].into();
-                } else {
-                    state.ambient_light = true;
-                    color.0 = [0.01; 3].into();
-                }
-            }
-            Some(VirtualKeyCode::D) => {
-                let mut lights = w.write_storage::<Light>();
-
-                if state.directional_light {
-                    state.directional_light = false;
-                    for light in (&mut lights).join() {
-                        if let Light::Directional(ref mut d) = *light {
-                            d.color = [0.0; 4].into();
-                        }
-                    }
-                } else {
-                    state.directional_light = true;
-                    for light in (&mut lights).join() {
-                        if let Light::Directional(ref mut d) = *light {
-                            d.color = [0.2; 4].into();
-                        }
-                    }
-                }
-            }
-            Some(VirtualKeyCode::P) => if state.point_light {
-                state.point_light = false;
-                state.light_color = [0.0; 4].into();
-            } else {
-                state.point_light = true;
-                state.light_color = [1.0; 4].into();
-            },
-            _ => (),
-        }
-        Trans::None
-    }
-
-    fn update(&mut self, data: StateData<GameData>) -> Trans<GameData<'a, 'b>> {
-        data.data.update(&data.world);
-        Trans::None
-    }
-}
-
-#[derive(Clone)]
-struct Assets {
-    cube: MeshHandle,
-    cone: MeshHandle,
-    lid: MeshHandle,
-    rectangle: MeshHandle,
-    teapot: MeshHandle,
-    red: Material,
-    white: Material,
-    logo: Material,
-    font: FontHandle,
-}
-
-fn load_assets(world: &World) -> (ProgressCounter, Assets) {
-    let mut progress = ProgressCounter::default();
-    let mesh_storage = world.read_resource();
-    let tex_storage = world.read_resource();
-    let font_storage = world.read_resource();
-    let mat_defaults = world.read_resource::<MaterialDefaults>();
-    let loader = world.read_resource::<Loader>();
-
-    let red = loader.load_from_data([1.0, 0.0, 0.0, 1.0].into(), &mut progress, &tex_storage);
-    let red = Material {
-        albedo: red,
-        ..mat_defaults.0.clone()
-    };
-
-    let white = loader.load_from_data([1.0, 1.0, 1.0, 1.0].into(), &mut progress, &tex_storage);
-    let white = Material {
-        albedo: white,
-        ..mat_defaults.0.clone()
-    };
-
-    let logo = Material {
-        albedo: loader.load(
-            "texture/logo.png",
-            PngFormat,
-            Default::default(),
-            &mut progress,
-            &tex_storage,
-        ),
-        ..mat_defaults.0.clone()
-    };
-
-    let cube = loader.load("mesh/cube.obj", ObjFormat, (), &mut progress, &mesh_storage);
-    let cone = loader.load("mesh/cone.obj", ObjFormat, (), &mut progress, &mesh_storage);
-    let lid = loader.load("mesh/lid.obj", ObjFormat, (), &mut progress, &mesh_storage);
-    let teapot = loader.load(
-        "mesh/teapot.obj",
-        ObjFormat,
-        (),
-        &mut progress,
-        &mesh_storage,
-    );
-    let rectangle = loader.load(
-        "mesh/rectangle.obj",
-        ObjFormat,
-        (),
-        &mut progress,
-        &mesh_storage,
-    );
-    let font = loader.load(
-        "font/square.ttf",
-        TtfFormat,
-        (),
-        &mut progress,
-        &font_storage,
-    );
-
-    (
-        progress,
-        Assets {
-            cube,
-            cone,
-            lid,
-            rectangle,
-            teapot,
-            red,
-            white,
-            logo,
-            font,
-        },
-    )
-}
-
-fn main() -> Result<(), Error> {
-    // Add our meshes directory to the asset loader.
-    let resources_directory = format!("{}/examples/assets", env!("CARGO_MANIFEST_DIR"));
-
-    let display_config_path = format!(
-        "{}/examples/renderable/resources/display_config.ron",
-        env!("CARGO_MANIFEST_DIR")
-    );
-
-    let game_data = GameDataBuilder::default()
-        .with(
-            UiLoaderSystem::<TextureFormat, FontFormat>::default(),
-            "",
-            &[],
-        )
-        .with::<ExampleSystem>(ExampleSystem::default(), "example_system", &[])
-        .with_bundle(TransformBundle::new().with_dep(&["example_system"]))?
-        .with_bundle(UiBundle::<String, String>::new())?
-        .with_bundle(HotReloadBundle::default())?
-        .with_bundle(FPSCounterBundle::default())?
-        .with_basic_renderer(display_config_path, DrawShaded::<PosNormTex>::new(), true)?
-        .with_bundle(InputBundle::<String, String>::new())?;
-    let mut game = Application::build(resources_directory, Loading::default())?.build(game_data)?;
-    game.run();
-    Ok(())
-}
-
-fn initialise_camera(world: &mut World) {
-    let mut local = Transform::default();
-    local.translation = Vector3::new(0., -20., 10.);
-    local.rotation = Quaternion::from_angle_x(Deg(75.)).into();
-    world
-        .create_entity()
-        .with(Camera::from(Projection::perspective(1.3, Deg(60.0))))
-        .with(local)
-        .with(GlobalTransform::default())
-        .build();
 }